Moved and re-referenced css files & updated App.js to class format.
